Split up TBitcoinNetwork::process()
If a function is too long to fit in one screen, then it's too long.

It's far more pleasant to read multiple discrete functions than a series
of functions hidden in a series of if()s.

This patch splits each of the message type handlers into their own
function within TBitcoinNetwork.  It is functionally identical (barring
an extra call), but is considerably more readable.