Make two classes for referencing transaction inputs and outputs
While they are actually identical in shape: one hash and one integer to
reference the input/output; for type safety it's better not to use the
same structure.

Given the complexity of what goes on in the transaction handling for
Bitcoin, it won't do any harm to be explicit about exactly what data
structure we're talking about when we return a reference.

This patch splits TCoinTransfer::sSplitReference into TInputReference
and TOutputReference and gives them a common base class.  That base
class further supplies isNull() by comparing the member variables with
the network-defined constants.

TCoinTransfer::getCreatorReference() and getClaimantReference() then
return the appropriate structure type.

A more appropriate name is also given to
TNetworkParameters::COINBASE_REFERENCE_HASH; which the rest of this
patch makes clear should be NULL_REFERENCE_HASH, since it is only
indirectly related to the idea of a coinbase (i.e. that a coinbase
transaction has a NULL reference for an input).