Rewrite TMessageFactory::receive() to deal with corrupt and multiple messages correctly
Sometimes we might get an incomplete message; sometimes we might get two
messages; sometimes we might get a whole message and then an incomplete
message.  Sometimes we might get junk (or a message we don't support)
and then a message.

TMessageFactory::receive() was handling these cases badly.  This patch
fixes it by repeatedly attempting a read starting at each byte in the
available buffer.  That means if it has ever gone out of sync with the
stream, it will find its way back, and that if there are multiple
messages available, they will be found too.