When the network parameters aren't known, pick the network using the message magic number
If the network parameters haven't been chosen (i.e. prodnet or testnet),
then we can look at the magic number in the first message we receive and
work out which network we are connected to by comparing against our
known network's magic numbers.

If the network parameters have been set, then we don't do the above, we
instead require the messages we receive to be for the network we've
already chosen.