Split TBitcoinScriptBase from TBitcoinScript to make non-streamed scripts possible
Scripts can come from a stream; but they can also be constructed programmatically.
This patch splits the stream-readable version (TBitcoinScript) from the
version that simply represents a script (TBitcoinScriptBase).

TBitcoinScriptBase holds the program, which is a live set of
TStackOperators; TBitcoinScript holds a set of TStackOperatorFromStreams
templates that will be matched against stream opcodes to construct a
program.  Finally a TBitcoinScript has a minimum acceptable version, but
TBitcoinScriptBase does not.  This is because it is the list of
template operators that represents the version, not the program itself.

The split also changes the way a script is loaded into
TBitcoinScriptBase; instead of calling append() multiple times, an array
of operators is passed.  This is to make it easy to construct a
TBitcoinScriptBase from a hard-coded script written in the source.