Remove set/get NetworkHash from TMessageBasedBlock
I made a mistake: the block message does not include the hash of the
block; so worrying about comparing the network hash with the calculated
hash is not a concern.  It can't be done.  The block is sent without a
hash, and we are expected to recalculate it from its contents.