Clear the cached hash when the message is changed in TMessageBasedBlock::updateFromMessage()
This would only be relevant if a single block is updated with a further
message.  Or updated by a message when it was previously updated in a
different way.  The cached hash has to be cleared ready for a new one to
be calculated.

This may be particularly relevant when we're talking about going from a
parent block who's hash we know from a reference, to a parent block
who's hash we calculated after its contents are subsequently received.