Set OP in TStackOperator_PUSH_N to the size of the given string
This lets us call the base class to write the opcode.  There is strictly
no difference, since the base class just writes the opcode byte, which
is now the same as the length which is all we were doing anyway; but
reuse of code avoids accidental bugs creeping in when changes are made
in base classes that aren't applied in child classes.