Calling TMessageFactory::answer(NULL) is the factory's opportunity to send unprompted messages
For example; we might want to send a message every five minutes.  There
is no received message that we would be answering.  Passing NULL to the
answer() member is used as a special case signal to allow the message
factory to send whatever it wants.

We then use that facility to allow TVersioningMessageFactory::answer()
to send the first 'version' message.