Add a TNetworkParameters parameter, GETDATA_MAX_INV
From the official client:

    else if (strCommand == "getdata")
    {
        vector<CInv> vInv;
        vRecv >> vInv;
        if (vInv.size() > 50000)
            return error("message getdata size() = %d", vInv.size());

I'm choosing not to sprinkle these arbitrary literals throughout the
code; instead giving them a name and making them a parameter in
TNetworkParameters.

This might be overkill, but even if they end up as, near as matters, constants,
we have a place to put them and their presence in the TNetworkParameters
class tells a reader that they are simply "chosen" numbers.