Add a self-connection check in TMessage_version_10600
From version 10600 onwards a 'nonce' has been included in the version
message.  This is a randomly chosen 32 bit number.  Given that a node
has no way of knowing it's own external IP and that another peer could
give it its own external IP as a potential peer, we need a way (other
than by IP) of checking that an connection isn't a self-connection.  We
do that by selecting a random 32-bit number at start up as a nonce,
then handing that over in the version message and then comparing the
nonce we receive with our own nonce.

The chance of two nodes picking the same nonce are small(ish, see the
birthday paradox) so we reject any peer with the same nonce as us, as it
is most likely us that is that peer.