`getblocks` and `getdata` implementations
A typical conversation between two peers might go like this:

  RX< getblocks
  TX> inv
  RX< getdata
  TX> block

The `getblocks` message contains (we presuppose) an inventory of chain
tips that the peer knows about.  It doesn't know which of these we
consider to be on the longest chain because it doesn't have the blocks
following them to be able to work that out.

We can work it out though, so we go though the inventory of block hashes
in the `getblocks` message and return children of that hash that are on
the longest chain.  We return those children in the form of an `inv`
message.

The peer then uses that `inv` to decide which blocks it wants and
requests them using getdata.

There is a programmed limit on how many items can be in the `inv`
response; so a hack is used when more are needed (probably during
initial block chain download).  A so-called "continuation hash" is
stored (in the peer object) and is checked when a `getdata` is received.
This continuation hash is the hash of the last block hash in an `inv`
that hit the hash limit.  When `getdata` finds a request for thatQ
particular block it instantly sends another `inv` with the hash tip of
the current longest chain.

The above is a description of what the official client does; and we'll
have to copy that.  It is manifestly wrong though.

The easier solution for bringing a peer up to speed would have been to
introduce a `getlongest` message, which would give the peer access to
the latest block.  Each block holds the parent hash, and it can then
easily request those parents for as long as it likes until what it
receives matches what it already has or until it reaches the genesis
block.  The convoluted mess that is this getblocks/getdata behaviour is
a result of traversing the chain forwards instead of backwards.
`getlongest` could even have been part of the `version` message (instead
of the fairly useless 'height' field that currently is sent).

Here's the icing though: `getheaders`, which is analogous to `getblocks`
and `getdata` combined (in that it sends headers back, but it does so
without needing an `inv` in between), doesn't use any continuation
mechanism at all, because it has no limiter, allowing 2000 more headers
than exist between the requested point and the longest chain tip.