TMessageBasedBlock::getHash() returns an invalid hash when no pool is set
It should never be the case that a block isn't in a pool, but we should
protect against it anyway.

The key result of this code is that a block not in a pool has no hash.
How can it?  It is the pool that tells it the network, and it is the
network that tells it what the hash function is.