Implement TBlockPool::getCommonAncestor()
Finds the block (if any) that is a common ancestor of the two given
blocks.  If they are both in the same chain, then the common ancestor
will be the eldest of the two.  If they are on different chains, the
common ancestor will be the branch point of those two chains.