mlock() private memory in TAutoClearAllocator
If a chunk of memory is private enough that it should be cleared before
freeing it, then it is important enough not to be flushed to the swap
file (where it is much easier for an attacker to look at it).  The
mlock() function tells the OS to do just that.

The official client's private allocator is so similar to ours that this
patch applies almost one-for-one.

https://github.com/bitcoin/bitcoin/commit/c1aacf0be347b10a6ab9bbce841e8127412bce41