Provide object-oriented call back class for TBitcoinNetwork events
Eventually we will have bitcoin-related events that an application using
this library will want to be informed of.  Instead of forcing a GUI to
poll us regularly, this patch provides the scaffolding of a call back
system.

This being in C++ rather than C though, we have no need to deal with the
untidyness of callback functions.  Instead we provide a base class with
virtual member functions for all the call backs, and TBitcoinNetwork
with a function for registering a callback object with it.

TBitcoinNetwork can then call the virtuals in that supplied object.

A user of the bitcoin library creates a child class of
TBitcoinEventObject and implements whichever of the callback members it
wants, keeping whatever additional pointers and variables it wants in
itself.