Add a static TCoinTransfer::NULL_REFERENCE to act as the reference-to-nowhere needed for coinbase transactions
A coinbase transaction has one input; that input is a reference to
nowhere (see TNetworkParameters::NULL_REFERENCE_HASH and
NULL_REFERENCE_INDEX).  However, we don't keep references themselves in
the TTransaction; we keep pointers to TCoinTransfer objects.  Therefore
when we make a coinbase transaction, we need an object to act as that
NULL reference.

This patch adds a static NULL_REFERENCE to act as that object.  The same
one will do for all coinbase transactions, so it is no problem being
static.