Add example TBitcoinEventObject child class
The default event object was replaced with a silent, do nothing, object.
This patch adds in an example event object that I can start moving all
the noise into.

All the noisy output should be moved to one place as that's the only way
I'll know that a GUI could reasonably replace a command line version.