Add TBitcoinAddress to hashtypes module
TBitcoinAddress is a special case of a base58 encoded number.  While the
number encodes and decodes as normal, the byte stream represents three
fields.  TBitcoinAddress gives access to those fields, and more
importantly allows them to be validated.

For convenience it also supports setting the key hash field from a
TEllipticCurveKey.