TBitcoinNetwork::receive_getdata() wasn't queuing the continuation 'inv' message
If the 'getblocks' responder wants to send more blocks than it is allow
by the network limits; then it sets a continuation hash.

That continuation hash is checked for in the 'getdata' responder so that
when the peer requests the last block that 'getblocks' was able to send
it; a new 'inv' is triggered offering the latest branch block.  This
should force the peer to perform a new 'getblocks' when it finds that it
doesn't have the parent of that latest branch block and the process
begins again.