[lib] Template parameters were wrong in TSecureByteArray definition
TSecureByteArray is a typedef that was being declared like this:

    typedef TByteArray_t<TAutoClearAllocator<allocator<unsigned char> > > TSecureByteArray;

The parameters for TAutoClearAllocator are defined like this:

    template<typename T, typename baseAllocator = std::allocator<T> >
    class TAutoClearAllocator : public baseAllocator

The allocator is the _second_ parameter to TAutoClearAllocator; not the
first.  The second will default to something reasonable given the type
of the storage required so the following works fine:

    typedef TByteArray_t<TAutoClearAllocator<unsigned char> > TSecureByteArray;

This makes it an <unsigned char> allocator.