[lib] Automatically calculate TNetworkParameters::MAX_MONEY and show how it's done
The 21 million bitcoin limit is not arbitrary, it is a result of the
selection of block reward and inflation period.  This patch shows how
the value is calculated and changes it to be calculated at runtime.

It makes no difference to the final outcome, but might be useful for
someone picking a different set of parameters -- it removes the
suggestion that MAX_MONEY is a free parameter, when it isn't.